refactor(test): split skill-injection-recording.test.ts for FR-006 cap (#1770)
PR feedback F-001: the new test file was 895 lines, exceeding the FR-006
500-line cap enforced by scripts/check-test-file-cap.sh (on main since PR #1802).
Split into 3 focused files, each under 500 lines, preserving all 30 tests:

- skill-injection-skip-and-none.test.ts  (249 lines, 9 tests)
  skip/no-op paths, SKILLS:none branch, error resilience
- skill-injection-qualified.test.ts      (451 lines, 15 tests)
  qualified-injection mechanics, attribution, threshold, reviewer forwarding
- skill-injection-roundtrip.test.ts      (359 lines, 6 tests)
  regression guard, SKILL_COMPLIANCE round-trip, gate handoff

Each file is self-contained (own helpers + imports) per repo convention.
Verified: 30/30 tests pass; cap check exits 0 (0 violations).
